backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2
psycopg2-binary==2.9.10
celery[redis]==5.4.0
django-celery-results==2.5.1
anthropic==0.40.0
duckdb==1.1.3
pandas==2.2.3
gunicorn==23.0.0
djangorestframework-simplejwt>=5.3
channels[daphne]>=4.0
channels-redis>=4.2

openpyxl>=3.1.5
pyarrow>=18.0.0

# Dev / test
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
